feat: Add optimization for output speed when `out` is not a terminal
In this commit, I added the `is_terminal` crate to optimize the output
speed in non-terminal environments for my Rust code. This was achieved
by creating a new `Out` struct and checking whether `stdout` is a
terminal or not. If it is not a terminal, then the `BufWriter` is used
for optimal performance.

The new code adds a `is_terminal` check and optimizes the output
accordingly while writing the results of the `get_download_info`
function to the `out` variable.

This feature improves the performance of the code when writing output to
non-terminal environments.